cron/requirements.txt
crontab==0.20.3
kafka-python==1.3.5
requests
json